LPS-57247 Move the process interface and registry util to a different package
