+++ modules/apps/blogs/blogs-service/src/com/liferay/blogs/lar/BlogsEntryExportImportContentProcessor.java
@@ -14,7 +14,7 @@
 
 package com.liferay.blogs.lar;
 
-import com.liferay.exportimport.api.ExportImportContentProcessor;
+import com.liferay.exportimport.content.processor.ExportImportContentProcessor;
 import com.liferay.exportimport.lar.BaseExportImportContentProcessor;
 import com.liferay.portal.kernel.log.Log;
 import com.liferay.portal.kernel.log.LogFactoryUtil;
