+++ modules/apps/blogs/blogs-service/src/com/liferay/blogs/lar/BlogsEntryStagedModelDataHandler.java
@@ -14,7 +14,7 @@
 
 package com.liferay.blogs.lar;
 
-import com.liferay.exportimport.api.ExportImportContentProcessor;
+import com.liferay.exportimport.content.processor.ExportImportContentProcessor;
 import com.liferay.exportimport.lar.BaseStagedModelDataHandler;
 import com.liferay.portal.kernel.dao.orm.QueryUtil;
 import com.liferay.portal.kernel.exception.PortalException;
