+++ modules/apps/export-import/export-import-api/bnd.bnd
@@ -2,6 +2,6 @@ Bundle-Name: Liferay Export Import API
 Bundle-SymbolicName: com.liferay.exportimport.api
 Bundle-Version: 1.0.0
 Export-Package:\
-	com.liferay.exportimport.portlet.preferences.processor,\
-	com.liferay.exportimport.api
+	com.liferay.exportimport.content.processor,\
+	com.liferay.exportimport.portlet.preferences.processor
 Include-Resource: classes
