+++ modules/apps/export-import/export-import-api/src/com/liferay/exportimport/content/processor/ExportImportContentProcessorRegistryUtil.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.exportimport.api;
+package com.liferay.exportimport.content.processor;
 
 import aQute.bnd.annotation.ProviderType;
