+++ modules/apps/wiki/wiki-service/src/com/liferay/wiki/lar/WikiPageExportImportContentProcessor.java
@@ -14,7 +14,7 @@
 
 package com.liferay.wiki.lar;
 
-import com.liferay.exportimport.api.ExportImportContentProcessor;
+import com.liferay.exportimport.content.processor.ExportImportContentProcessor;
 import com.liferay.exportimport.lar.BaseExportImportContentProcessor;
 import com.liferay.portal.kernel.log.Log;
 import com.liferay.portal.kernel.log.LogFactoryUtil;
