+++ modules/apps/wiki/wiki-service/src/com/liferay/wiki/lar/WikiPageStagedModelDataHandler.java
@@ -14,7 +14,7 @@
 
 package com.liferay.wiki.lar;
 
-import com.liferay.exportimport.api.ExportImportContentProcessor;
+import com.liferay.exportimport.content.processor.ExportImportContentProcessor;
 import com.liferay.exportimport.lar.BaseStagedModelDataHandler;
 import com.liferay.portal.kernel.dao.orm.QueryUtil;
 import com.liferay.portal.kernel.exception.PortalException;
