+++ portal-service/src/com/liferay/portlet/exportimport/lar/ExportImportHelperUtil.java
@@ -322,7 +322,7 @@ public class ExportImportHelperUtil {
 
 	/**
 	 * @deprecated As of 7.0.0, replaced by
-	 *             com.liferay.exportimport.api.ExportImportContentProcessorr#replaceExportContentReferences(
+	 *             com.liferay.exportimport.content.processor.ExportImportContentProcessorr#replaceExportContentReferences(
 	 *             PortletDataContext, StagedModel, String, boolean, boolean) in
 	 *             the export-import-api module
 	 */
@@ -340,7 +340,7 @@ public class ExportImportHelperUtil {
 
 	/**
 	 * @deprecated As of 7.0.0, replaced by
-	 *             com.liferay.exportimport.api.ExportImportContentProcessor#replaceExportContentReferences(
+	 *             com.liferay.exportimport.content.processor.ExportImportContentProcessor#replaceExportContentReferences(
 	 *             PortletDataContext, StagedModel, String, boolean, boolean) in
 	 *             the export-import-api module
 	 */
@@ -358,7 +358,7 @@ public class ExportImportHelperUtil {
 
 	/**
 	 * @deprecated As of 7.0.0, replaced by
-	 *             com.liferay.exportimport.api.ExportImportContentProcessorr#replaceExportContentReferences(
+	 *             com.liferay.exportimport.content.processor.ExportImportContentProcessorr#replaceExportContentReferences(
 	 *             PortletDataContext, StagedModel, String, boolean, boolean) in
 	 *             the export-import-api module
 	 */
@@ -459,7 +459,7 @@ public class ExportImportHelperUtil {
 
 	/**
 	 * @deprecated As of 7.0.0, replaced by
-	 *             com.liferay.exportimport.api.ExportImportContentProcessor#replaceImportContentReferences(
+	 *             com.liferay.exportimport.content.processor.ExportImportContentProcessor#replaceImportContentReferences(
 	 *             PortletDataContext, StagedModel, String) in the
 	 *             export-import-api module
 	 */
@@ -476,7 +476,7 @@ public class ExportImportHelperUtil {
 
 	/**
 	 * @deprecated As of 7.0.0, replaced by
-	 *             com.liferay.exportimport.api.ExportImportContentProcessor#replaceImportContentReferences(
+	 *             com.liferay.exportimport.content.processor.ExportImportContentProcessor#replaceImportContentReferences(
 	 *             PortletDataContext, StagedModel, String) in the
 	 *             export-import-api module
 	 */
